src/api-server/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
cachetools==4.0.0         # via google-auth
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via flask
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via networkx
dnspython==1.16.0         # via -r requirements.in
fasttext==0.9.1           # via -r requirements.in
flask==1.0                # via -r requirements.in, flask-cors
flask_cors==3.0.0         # via -r requirements.in
google-api-core[grpc]==1.16.0  # via google-cloud-vision
google-auth==1.12.0       # via google-api-core
google-cloud-vision==0.34.0  # via -r requirements.in
google-cloud==0.34.0      # via -r requirements.in
googleapis-common-protos==1.51.0  # via google-api-core
grpcio==1.27.2            # via google-api-core
idna==2.9                 # via requests
imageio==2.8.0            # via scikit-image
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via flask
kiwisolver==1.1.0         # via matplotlib
langdetect==1.0.8         # via -r requirements.in
markupsafe==1.1.1         # via jinja2
matplotlib==3.2.1         # via scikit-image
networkx==2.4             # via scikit-image
nltk==3.4.5               # via textblob
numpy==1.18.2             # via fasttext, imageio, matplotlib, pywavelets, scipy, torch, torchvision
pillow==7.1.0             # via imageio, scikit-image, torchvision
protobuf==3.11.3          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pybind11==2.4.3           # via fasttext
pymongo==3.10.1           # via -r requirements.in
pyparsing==2.4.6          # via matplotlib
python-dateutil==2.8.1    # via matplotlib
pytz==2019.3              # via google-api-core
pywavelets==1.1.1         # via scikit-image
requests==2.23.0          # via google-api-core
rsa==4.0                  # via google-auth
scikit-image==0.15.0      # via -r requirements.in
scipy==1.4.1              # via scikit-image
six==1.14.0               # via cycler, flask-cors, google-api-core, google-auth, grpcio, langdetect, nltk, protobuf, python-dateutil, torchvision
textblob==0.15.3          # via -r requirements.in
torch==1.2.0              # via -r requirements.in, torchvision
torchvision==0.4.0        # via -r requirements.in
tqdm==4.44.0              # via -r requirements.in
urllib3==1.25.8           # via requests
werkzeug==1.0.0           # via flask
wget==3.2
pika==1.1.0
opencv-contrib-python==4.4.0.42
elasticsearch

# The following packages are considered to be unsafe in a requirements file:
# setuptools
